Abstracts RandomValueGenerator fully ready for use by other agents
